use DOM element from state
The component will re-render when the DOM node is set in state and the
effect will re-run with the updated DOM node. When using a ref, the
component doesn't rerender, which means that if the ref isn't filled in
yet (because of a transition for example), then it won't re-run the
effect once it is filled in.