Fix `Combobox` `virtual` mode types for `multiple`
When multiple is specified we expect TValue to already be an array, so lets just
ensure its an array and use it as the type for virtual.options. Similarly the
disabled callback gets passed a TValue, but when multiple is true it is passed
something of type element of TValue.

Behaviour in javascript lines up with this fine, just a type level fix afaik.