use local state for transitions
This now relies on local state for transitions so that when the local
state changes, the current component can re-render on its own already.
Relying on state from the parent via context, requires the full
component to re-render first which is not efficient enough for this
case.

We still need that more global state for cases where we want to
reference an element in another component (e.g.: reference the panel id
in a button's aria-controls attribute)