rely on `node.getAnimations()` to know when transitions are done
Up until now, we've tried to do this ourselves by listening to the
correct events. The new `node.getAnimations()` API is a much simpler API
to use.

The only requirement is that we call `node.getAnimations()` in a
`requestAnimationFrame` so that the browser can flush all the changes.
We couldn't do this before, because we needed to setup the event
listeners to prevent race conditions.

Now there are no race conditions, in fact, if all transitions already
complete before we can call the `waitForTransition`, then the
`node.getAnimations()` list will be empty and we can call the `done()`
function.

The `Element.prototype.getAnimations` has been available in browsers
since mid 2020, but at the time it was too new to use. Now seems like a
safe time to use this.

See: https://developer.mozilla.org/en-US/docs/Web/API/Element/getAnimations#browser_compatibility