update failing test
The order slightly changed here, but it looks more correct now as well.
The `child-2-1` and `child-2-2` have the same duration/delay, so the
order that they are rendered in seems to be the correct order.

That said, these DOM nodes are on the same level (siblings), so for
these callbacks we are testing here the order doesn't really matter. The
order that _does_ matter is that the `after*` events of a parent fire
_after_ the `after*` events of the children, which it still does.